Fix: CoreData 기본 세팅위치 변경
AppDelegate에서 기본설정하는 작업을 SceneDelegate에서 설정하도록 수정하였습니다.
