+++ SueobHaro/SueobHaro/ViewController.swift
@@ -6,12 +6,17 @@
 //
 
 import UIKit
+import CoreData
 
 class ViewController: UIViewController {
+    
+    var container: NSPersistentContainer!
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view.
+        guard container != nil else { fatalError("This view needs a persistent container.") }
+        
+        
     }
